pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "py-dir-fake-users-from-channel"
version = "0.1.0"
description = "Telegram channel fake users scanner and remover"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "telethon>=1.34.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.0",
    "pandas>=2.0.0",
    "tabulate>=0.9.0",
    "asyncio-throttle>=1.0.2",
    "aiofiles>=23.2.1",
    "openpyxl>=3.1.0",
    "aiosqlite>=0.21.0",
]
authors = [
    {name = "Telegram Scanner Team"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Communications :: Chat",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Homepage = "https://github.com/yourusername/py-dir-fake-users-from-channel"
Repository = "https://github.com/yourusername/py-dir-fake-users-from-channel.git"
Issues = "https://github.com/yourusername/py-dir-fake-users-from-channel/issues"

[project.scripts]
telegram-scanner = "telegram_scanner.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
telegram_scanner = ["py.typed"]
